# backend/internal/platform/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Verified Dating Gateway API
  version: 1.0.0
  description: |
    Mobile-facing API exposed by API Gateway + Mobile BFF.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/billing/subscribe:
    post:
      summary: Subscribe user to billing plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, plan_id]
              properties:
                user_id:
                  type: string
                plan_id:
                  type: string
                billing_cycle:
                  type: string
      responses:
        '200':
          description: Billing subscription accepted
...